cleanup customization example
