+++ examples/pages/Customization/data.js
@@ -1,0 +1,31 @@
+export const videos = [
+  {
+    poster: 'https://peach.blender.org/wp-content/uploads/bbb-splash.png',
+    urls: [
+      {
+        type: 'video/mp4',
+        src: 'http://download.blender.org/peach/trailer/trailer_480p.mov',
+      },
+      {
+        type: 'video/ogg',
+        src: 'http://download.blender.org/peach/trailer/trailer_400p.ogg',
+      },
+    ],
+  },
+  {
+    poster:
+      'http://www.gigareel.com/wp-content/uploads/2016/04/i_love_sintel_by_nash88-d2zkxw5.jpg',
+    urls: [
+      {
+        type: 'video/mp4',
+        src:
+          'https://download.blender.org/durian/trailer/sintel_trailer-480p.mp4',
+      },
+      {
+        type: 'video/ogg',
+        src:
+          'https://download.blender.org/durian/trailer/sintel_trailer-480p.ogv',
+      },
+    ],
+  },
+];
